icees/requirements.txt
pandas==2.0.3
openai==1.10.0
python-dotenv==1.0.1

